Create a statusFactory that does formatting on various properties of the status object before creating a status
